Refactor various TypeScript files for improved type handling and code clarity
- Updated toolProviderFactory.ts to ensure proper type casting in get_tools method.
- Modified getCaseFileDocument.ts to enhance type safety in mapping documents.
- Refined manager-body.ts for better readability and structure in email processing logic.
- Improved logger core.ts to handle logger callbacks more efficiently.
- Adjusted user-keys.ts for consistent async handling of public key retrieval.
- Removed outdated dependency from package.json.
- Added comprehensive unit tests for Call-to-Action API route, covering GET and POST endpoints, pagination, filtering, and error handling.
- Created selectForGridV2.ts as a placeholder for future query helper implementations.
