+++ web-ui/__tests__/components/ai/chat-panel/chat-panel-dock.test.tsx
@@ -26,22 +26,9 @@ jest.mock('@/lib/ai/core', () => ({
   isAnnotatedRetryMessage: () => false,
 }));
 
-jest.mock('@/lib/logger', () => ({
-  log: () => {},
-}));
-
 jest.mock('@/lib/components/ai/chat-fetch-wrapper', () => ({
     useChatFetchWrapper: jest.fn(() => ({ chatFetch: jest.fn() })),
 }));
-
-jest.mock('@/instrument/browser', () => ({
-  getReactPlugin: () => ({}),
-}));
-
-jest.mock('@microsoft/applicationinsights-react-js', () => ({
-  withAITracking: (plugin: any, Component: any) => Component,
-}));
-
