+++ web-ui/components/chat/chat-message-display.tsx
@@ -84,12 +84,10 @@ export const ChatMessageDisplay: React.FC<ChatMessageDisplayProps> = ({
           </IconButton>
         )}
       </Box>
-
       {/* Message Content */}
       <Typography variant="body2" sx={{ whiteSpace: 'pre-wrap', mb: 1 }}>
         {message.content || '<no content>'}
       </Typography>
-
       {/* Optimized Content (if different from regular content) */}
       {message.optimizedContent && message.optimizedContent !== message.content && (
         <Accordion>
@@ -103,7 +101,6 @@ export const ChatMessageDisplay: React.FC<ChatMessageDisplayProps> = ({
           </AccordionDetails>
         </Accordion>
       )}
-
       {/* Metadata Section */}
       {showMetadata && (
         <Collapse in={metadataExpanded}>
@@ -113,37 +110,37 @@ export const ChatMessageDisplay: React.FC<ChatMessageDisplayProps> = ({
               Message Metadata
             </Typography>
             <Grid container spacing={1}>
-              <Grid xs={6}>
+              <Grid size={6}>
                 <Typography variant="caption" display="block">
                   Message ID: {message.messageId}
                 </Typography>
               </Grid>
-              <Grid xs={6}>
+              <Grid size={6}>
                 <Typography variant="caption" display="block">
                   Order: {message.messageOrder}
                 </Typography>
               </Grid>
               {message.providerId && (
-                <Grid xs={6}>
+                <Grid size={6}>
                   <Typography variant="caption" display="block">
                     Provider: {message.providerId}
                   </Typography>
                 </Grid>
               )}
-              <Grid xs={6}>
+              <Grid size={6}>
                 <Typography variant="caption" display="block">
                   Status ID: {message.statusId}
                 </Typography>
               </Grid>
               {message.toolInstanceId && (
-                <Grid xs={12}>
+                <Grid size={12}>
                   <Typography variant="caption" display="block">
                     Tool Instance: {message.toolInstanceId}
                   </Typography>
                 </Grid>
               )}
               {message.functionCall && (
-                <Grid xs={12}>
+                <Grid size={12}>
                   <Typography variant="caption" display="block">
                     Function Call:
                   </Typography>
