+++ web-ui/package.json
@@ -131,7 +131,6 @@
   "overrides": {
     "uri-js": "npm:uri-js-replace",
     "psl": "npm:@imput/psl@2.0.4",
-    "@auth/core": "0.39.1",
     "data-urls@3.*": {
       "whatwg-url": "12.0.1"
     },
